+++ lib/index.ts
@@ -5,29 +5,32 @@ export { ScrapQuery, ScrapSelector, scrap } from './scrapper';
 export { attrCreator as attr } from './selectors/attr';
 export { existsCreator as exists } from './selectors/exists';
 export { htmlCreator as html } from './selectors/html';
-export { listCreator as list } from './selectors/list';
 export { textCreator as text } from './selectors/text';
 export { selectCreator as select } from "./selectors/select";
 export { countCreator as count } from "./selectors/count";
 export { linkCreator as link } from "./selectors/link";
 
+export { listCreator as List } from './controls/list';
+export { ifCreator as If } from './controls/if';
+
 // Import
 import { attrCreator as attr } from './selectors/attr';
 import { existsCreator as exists } from './selectors/exists';
 import { htmlCreator as html } from './selectors/html';
-import { listCreator as list } from './selectors/list';
 import { textCreator as text } from './selectors/text';
 import { selectCreator as select } from "./selectors/select";
-import { ifCreator as If } from './controls/if';
 import { countCreator as count } from "./selectors/count";
 import { linkCreator as link } from "./selectors/link";
+
+import { listCreator as List } from './controls/list';
+import { ifCreator as If } from './controls/if';
 
 export const Q = {
 	If,
 	attr,
 	exists,
 	html,
-	list,
+	List,
 	text,
 	select,
 	count,