+++ test/data.ts
@@ -1,0 +1,14 @@
+export const html = `
+    <div>
+        <h1 class="title">Hello</h1>
+        <ul>
+            <li><span>Guten Tag</span></li>
+            <li><span class="msg">Ciao</span></li>
+            <li><span>Bonjour</span></li>
+        </ul>
+        <a href="/read-more">read more ...</a>
+        <div class="footer" data-extra="footer">
+            Best translation since 2000
+        </div>
+    </div>
+`;