+++ test/exhaustive/agescx.test.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from "fs";
-import { scrap, list, text, link, html } from "../../lib";
+import { scrap, List, text, link, html } from "../../lib";
 
 const shtml = readFileSync('./test/exhaustive/agescx.html').toString();
 
@@ -7,10 +7,10 @@ describe('agescx documentation', () => {
 
     it('should get all navigation items from agescx', () => {
         const result = scrap(shtml, {
-            navs: list('ul.navbar-nav>li:not(.disabled)', {
+            navs: List('ul.navbar-nav>li:not(.disabled)', {
                 text: text('li > a'),
                 link: link('li > a'),
-                submenu: list('ul.dropdown-menu>li', {
+                submenu: List('ul.dropdown-menu>li', {
                     text: text('li > a'),
                     link: link('li > a')
                 })
@@ -18,7 +18,7 @@ describe('agescx documentation', () => {
             title: text('h1'),
             content: html('div[role="main"]')
         });
-        
+
         expect(result.navs.length).toBe(9);
         expect(result.navs[2].submenu.length).toBe(2);
         expect(result.navs[2].submenu[0].text).toBe('Adding new unit');